requirements.txt
logzero==1.7.0
PyP100==0.0.19
pyyaml==5.3.1
cerberus==1.3.4
requests==2.28.1